1213 임현정 Refactor: 구글맵 리팩토링 및 컴포넌트 분리
